+++ src/kanban/node_layout.rs
@@ -392,6 +392,12 @@ impl NodeLayout {
             ));
         }
     }
+    pub fn scroll_to(&mut self, id: KanbanId) {
+        let target_rect = self.sense_regions.iter().find(|x| x.0 == id);
+        if let Some(target_rect) = target_rect {
+            self.scene_rect.set_center(target_rect.1.center());
+        }
+    }
     pub fn show(
         &mut self,
         _document: &KanbanDocument,